requirements.txt
boto3
flask
gunicorn
Ipython
python-dotenv
langchain
langchain-community
langchain-google-genai